+++ lib/details/controller/one_crypto_provider.dart
@@ -1,7 +1,7 @@
-import 'package:crypto/details/controller/variation_notifier.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import '../../shared/models/crypto_model.dart';
+import 'details_crypto_notifier.dart';
 
 var oneCryptoProvider =
     StateNotifierProvider<DetailsCryptoNotifier, CryptoModel>(
