+++ lib/details/controller/variation_notifier.dart
@@ -1,5 +1,4 @@
 import 'package:crypto/shared/models/crypto_model.dart';
-import 'package:decimal/decimal.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 var variationProvider = StateNotifierProvider<VariationNotifier, double>(
@@ -17,27 +16,5 @@ class VariationNotifier extends StateNotifier<double> {
             1) *
         100;
     state = variation;
-  }
-}
-
-class DetailsCryptoNotifier extends StateNotifier<CryptoModel> {
-  DetailsCryptoNotifier()
-      : super(
-          CryptoModel(
-            id: 0,
-            name: '',
-            abbreviation: '',
-            howMuchUserHave: Decimal.parse('1'),
-            quantity: Decimal.parse('1'),
-            variation: 1.3,
-            currentPrice: Decimal.parse('1'),
-            iconImage: '1',
-            priceInOne: [Decimal.parse('1')],
-            priceInNinety: [Decimal.parse('1')],
-          ),
-        );
-
-  void changeDetailsCrypto(CryptoModel model) {
-    state = model;
   }
 }
