+++ lib/main.dart
@@ -1,24 +1,7 @@
-import 'package:crypto/shared/routes/app_routes.dart';
+import 'package:crypto/app_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 void main() {
-  runApp(const ProviderScope(child: MyApp()));
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Crypto',
-      theme: ThemeData(
-        primarySwatch: Colors.red,
-      ),
-      debugShowCheckedModeBanner: false,
-      routes: appRoutes,
-      initialRoute: '/portifolio',
-    );
-  }
+  runApp(const ProviderScope(child: AppWidget()));
 }
