+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   intl: ^0.17.0
   fl_chart: ^0.55.1
   decimal: ^2.3.0
+  device_preview: ^1.1.0
  
 
 dev_dependencies:
@@ -85,6 +86,7 @@ flutter:
     - family: Montserrat
       fonts:
         - asset: assets\fonts\Montserrat\static\Montserrat-Bold.ttf
+        - asset: assets\fonts\Montserrat\static\Montserrat-Regular.ttf
           weight: 700
         # - asset: assets/fonts/Montserrat/Montserrat-Italic-VariableFont_wght.ttf
         #   style: italic
